processor_service/requirements.txt
Flask==2.3.2
google-cloud-storage==2.11.0
google-cloud-firestore==2.11.0
google-cloud-aiplatform==1.30.0
google-api-core
gunicorn==20.1.0
pymupdf
